Don't store the cost and heuristic.
We don't need to store the cost and heuristic of nodes, as we only ever care
about the sum of the two values, which we can always infer from the node's
offset in the todo list. This saves 8 bytes per entry in the todo list and gives
us a 5% speed-up overall. More importantly it saves a non-trivial amount of
memory.

Whilst here, drop the pretence that we care about anything other than u32s. This
also lets us drop a try_from.