When we've found a success node, there's no point exploring higher cost nodes.
When we've found a success node, we still need to explore all reduces and shifts,
because both are zero cost, and may lead us to new success nodes. If the
heuristic is > 0, then we also need to keep exploring inserts and deletes.
However, if the heuristic == 0, then there's no point in exploring inserts and
deletes because, by definition, they will increase the cost beyond that of the
success node.

All of this rests on the long-standing assumption that term_cost|x| > 0 for
all x. This assumption is now enforced in this commit.

Although obviously highly dependent on one's use-case, this gives a 20-25%
performance increase on my standard examples.