It really doesn't make any difference if we parse further than we expected.
If we parse a bit further, it simply means that we've got a good repair:
throwing it away is perverse. This conditional therefore seems entirely
wrong-headed in retrospect and was probably based on a mis-reading of
Corchuelo's paper. It's not that likely to have made much difference in
practise, but it's best removed.