Merge compatible repairs in the Corchuelo recoverer.
Consider a grammar such as:

  S: T 'c' 'd';
  T: 'a' | 'b';

and assume we've got the input "d". We'll then find two repair sequences:

  Insert a, c
  Insert b, c

because T can generate two alternatives. While it's very helpful for the user to
see both possibilities, it's annoying from the search's perspective as both are
equivalent: we end up in the same place whether we insert 'a' or 'b'. However,
carrying both around doubles the search space (at least temporarily), causing us
to do a lot of extra work.

This commit merges repairs when they're compatible, so that we only have to
search for one of the two above repair sequences, only materialising both
if/after they turn out to be successful. In other words, we still find all the
same repair sequences as we did before, but have to search fewer nodes in order
to do so.

The first thing we need to do is define a notion of compatible search nodes. Two
search nodes are compatible iff:

  * Their la_idxs are identical.
  * Their pstacks are identical.
  * Their repairs are compatible. Two repair sequences are compatible if they
    end in the same number of shifts. [Ending in zero shifts is fine.]

To make that clearer, here are some compatible repair sequences:

  [Insert a, Shift] [Insert b, Shift]
  [Insert a] [Insert b]
  [Delete] [Insert a]
  [Delete, Shift, Shift] [Insert a, Shift, Shift]

Here are some incompatible repair sequences:

  [Insert a, Shift] [Insert b, Delete]
  [Insert a, Shift, Shift] [Insert a, Shift]

In order to actually merge things in, we have to do a bit of major surgery on
some parts of the code. First we say that search nodes have a cactus of
RepairMerges. Each RepairMerge is either a single Repair or a Merge of a Repair
and a set of Cactus<RepairMerge>s. When we add a node to the todo list, we look
to see if there is already a compatible node present; if so, we merge the new
node into the old node.

Imagine we have a node with repairs "[Insert a, Shift]" in our todo list and we
then merge a compatible node with repairs "[Insert b, Shift]" in. We will end up
with a representation along the lines of:

  repairs: [RepairMerge::Merge(a, [RepairMerge::Repair(b),
                                   RepairMerge::Repair(Shift)]),
            RepairMerge::Repair(Shift)]

As this shows, we can end up with trees of merges.

*Huge* numbers of nodes are compatible, however most nodes are unproductive
(i.e. they have no descendants). In other words, it's easy to merge things in
that are never used. The problem then becomes how to merge things in fast enough
that the cost of merging pays off. This took me quite a while to get right and
comes in two parts. First, we represent the set of merged repairs as a Cactus:
this plays nice with malloc, which ends up with lots of similarly sized objects
in memory. Second, the todo list becomes an IndexMap (i.e. an ordered map) from
"original" to "merged" nodes. The key for an element never changes, but as we
merge nodes in, the value changes. We can do this without memory allocation.

The good news is that this has a pretty significant impact on performance,
making bigger examples almost exactly 2x faster.