Add a recovery timeout.
Corchuelo have heuristics to know "when to stop" -- unfortunately they don't
work well in edge cases as they're based on the number of tokens examined. For
example, on the Java grammar, it's possible to construct inputs that have
massive search spaces, even with only a few tokens involved, and that take tens
of minutes to complete.

This commit adds a simple timeout facility to recoverers: after a fixed amount
of time (0.5s) the recoverer simply gives up. This is, in one way, very dumb,
but it deals well with the user expectation that recovery won't continue
indefinitely. The check is put into the "neighbours" function, which is expected
to regularly check the timeout. This means that we check pretty often, but not
so often as to be ridiculous. It also keeps the API fairly simple.