Add a test derived from Cerecke's thesis.
The example is one that Cerecke's LoopLimit check cannot repair. As the comment
in the code says, the repair given in the thesis contains a typo: it says 6 "a"s
are needed for a correct repair, but 7 "a"s are actually needed.

In order to make this test work, add some plumbing so that tests can specify
terminal costs.