Another test from Cerecke's thesis.
This test is one that shouldn't work, but does: because we currently derive
repairs from the stategraph, we can find repairs that result from non-LR1 parts
of the grammar. In English, this means that we can generate repairs (because the
stategraph is pre-conflict resolution) that we then can't parse (because of
conflict resolution).

This should be fixed.