Speed up state_actions().
A reasonable rule of thumb is that the state table is about 90% empty.
state_actions thus did vast numbers of fruitless hashmap lookups.

This commit adds a simple bitvector which records which entries in the
statetable have entries in. We can thus quickly iterate over this without having
to do any hashmap lookups at all.