Allow productions to appear "out of order".
This is a feature of Yacc that I simply hadn't considered. This grammar is
perfectly valid:

  A: 'a';
  B: 'b';
  A: 'c';

Previously we treated this as equivalent to:

  A: 'a'
   | 'c';
  B: 'b';

However this reordering is not safe in the presence of conflicts. Cerecke's PhD
thesis has this example on p54 which shows the issue:

  S: A 'c' 'd'
   | B 'c' 'e';
  A: 'a';
  B: 'a'
   | 'b';
  A: 'b';

If this CFG is later translated into an LR grammar, there will be conflicts
between:

  A: 'a'; and B: 'a'; [by Yacc's rules, the former is chosen]
  B: 'b'; and A: 'b'; [by Yacc's rules, the former is chosen]

However we silently rewrote this to:

  S: A 'c' 'd'
   | B 'c' 'e';
  A: 'a'
   | 'b';
  B: 'a'
   | 'b';

Meaning that A: 'a' was erroneously chosen instead of B: 'a'.

This commit preserves the ordering of "out of order" productions. It does this
by changing the representation of rules in the ASTs: rather than mapping rules
to productions, we map them to production indices. A rule may have
non-consecutive production indices.

The changes to non-test code are relatively simple. I've also taken the
opportunity to rid ourselves of the unmaintained linked-hash-map crate and move
to indexmap.

There is however a lot of churn to tests. Since so many needed changing, I've
also revisited them a little bit: the tests are now far more idiomatic Rust, and
I've removed one redundant test.