Add helper functions to make error recovery faster.
lrpar's error recovery functions spend a long time querying the statetable: that
means lots of hash table lookups which, cumulatively, become very costly. This
commit enables the recoverers to reduce the number of lookups.

There are a few related things in here:

  1) Use the new Range facility in Vob's iter_set_bits to speed up the existing
     state_actions function.

  2) Add a new helper function state_shifts which tells users only which entries
     are shifts (i.e. it won't tell you where reduces or accepts are). If you
     only care about searching for shifts, this saves a lot of hashtable
     lookups.

  3) Add a more esoteric helper function "core_reduces" which is best explained
     by the comment in the commit. This is specifically useful for the MF
     recoverer as it removes its dependency on the stategraph.

These make a big difference to the recoverers.

[FWIW, I also tried a very different approach, using an IndexMap which
represented an ordered sparse table, and a separate list of where each state's
entries were in that IndexMap. Perhaps because an IndexMap costs an extra
lookup, I couldn't measure any difference when using this approach. It also uses
a bit more memory, so I ended up marginally preferring the approach in this
commit.]