Have the MF recoverer use the statetable and not the stategraph for guidance.
Cerecke has an odd, but compelling example, which shows the dangers of using the
stategraph to guide repairs (as seen in the previous commit). This commit uses
some of the recent support in lrtable to allow the MF recoverer to purely use
the statetable for guidance. This means that we now find only repairs which can
be parsed, fixing the test derived from Cerecke's thesis.