Include ranking repairs in the timeout.
If we have lots of repair sequences, ranking them all can be slow; it therefore
makes sense to include this in our timeout. However, we don't make this a fatal
error (in the sense that we say "couldn't find any repairs"): we just sort
repair sequences by whatever we've managed to do thus far. In other words, we'll
still sometimes have done enough ranking to be useful; and we won't ever be
worse than non-deterministically selecting a repair sequence as the top one.