Remove unsafe optimisation.
I originally thought that:

  Insert a, Shift a
  Shift a, Insert a

were equivalent, but they're not in the face of compatible merges: there's an
asymmetry that can't be resolved without checking the user's input, which we
can't do in our implementation (because the compatibility check is done as part
of an IndexMap's natural behaviour). It's therefore best to remove this
optimisation, which is a lot less powerful than merging compatible nodes.