Be more careful with compatible repairs.
I forgot that when checking for compatible repairs we need to take into account
this:

  // We follow Corcheulo et al.'s suggestions and never follow Deletes with
  // Inserts.

For example with the MF recoverer and this Java 1.5 program:

  class C {
    void x(T ... y) { }
  }

before this patch we would find 4 repair sequences:

  Insert "IDENTIFIER", Delete ".", Delete "."
  Delete ".", Delete ".", Keep "y", Insert "IDENTIFIER"
  Insert "IDENTIFIER", Delete ".", Keep ".", Keep "y", Insert "IDENTIFIER"
  Insert "IDENTIFIER", Keep ".", Delete ".", Keep "y", Insert "IDENTIFIER"

The solution is fairly simple: we add an additional clause that if one repair
sequence ends with a Delete the other must too. That allows us to find a further
4 repair sequences:

  Insert "IDENTIFIER", Keep ".", Insert "IDENTIFIER", Delete "."
  Insert "IDENTIFIER", Delete ".", Keep ".", Insert "IDENTIFIER"
  Insert "IDENTIFIER", Keep ".", Insert "IDENTIFIER", Keep ".", Insert "IDENTIFIER"
  Insert "IDENTIFIER", Keep ".", Insert "IDENTIFIER", Keep ".", Keep "y", Insert "IDENTIFIER"

While here, tidy up the compatibility check to make it easier to read.