Sort repairs consistently, as far as we are able.
Previously, two equally ranked repair sequences, of the same cost, and the same
length would appear in a random order in the list of repair sequences. This can
be frustrating when comparing error recovery algorithms, where one repair
sequence is (far further forward in the file) better than another.

This commit reduces the randomness by sorting on the ParseRepair struct itself
as a third case, though it can't avoid it completely. For example, terminals are
non-deterministically assigned different numbers on each grammar generation, so
sometimes "Insert a" will be "InsertTerm(0)" and sometimes "InsertTerm(1)" and
fixing that will involve a non-trivial API fix.