Remove the PORTION_THRESHOLD limit.
The time-based limit we now have is far more effective and less confusing, so
simply remove this entirely. It also had some weird, unintended, effects. On the
expression grammar, we would find non-minimal cost repair sequences for inputs
that were on the PORTION_THRESHOLD limit (e.g. "(+++++)") because we would stop
searching nodes of >= PORTION_THRESHOLD for neighbours, but still see if they
were success nodes.