Try only one member of a compatible merge for how far it can parse.
When we've collected repair sequences, we then see how far they let us parse,
and select only the best from those. However, previously we flattened all repair
sequences and then tried them. If two configurations are compatible, then the
outcome is inevitably the same. So we can save ourselves quite a bit of work by
only trying one member of a compatible merge.

As luck would have it, this also makes the code a bit simpler.