Be less strict about rule names.
Previously we forced rule names to be "standard" identifiers. But this means
that you couldn't write rules such as:

  \\+ +

which is really rather annoying. This commit simply removes this restriction:
in essence, all the non-whitespace chars after the last space in the line become
the rule name. So the above says "the regular expression \\+ has the name +".

An interesting question is: why was this restriction in place? Well, it
more-or-less matches lex, in the sense that Lex would write the above as:

  \\+ { return PLUS; }

Where PLUS is a C-level identifier shared between yacc and lex. Because this is
annoying to use in grammars, yacc allows you to specify single-character string
literals in grammars:

  expr: expr '+' expr;

and so on. In such cases yacc does some lexing of its own before calling lex
(via the yylex() function)! That just seems backwards (or forwards, or whatever
the opposite of the "right way to do it" is) to me.

My conclusion: we ended up following the restrictions of a weird hack that
relied on C identifiers. Ultimately, we can rid ourselves of this problem and
still remain almost 100% lex/yacc compatible. The only thing is that if someone
uses string literals in their yacc grammar, they'll have to define a token in
their lex file. This seems like a very minor imposition.