Rule names now need to be quoted.
We opened up a can of worms (which, admittedly, we should have addressed anyway)
with 7b1e3056. We allowed things like this:

  - -

i.e. a rule with the regex "-" (LHS) and the name "-" (RHS). That works well for
nearly all cases, *except* we inherited from lex this idiom:

  [ \t\n\r]+ ;

where a rule name of ";" means "don't generate any tokens for the matched
tests." This raises an obvious problem for this rule:

  ; ;

which a user would probably think means "match semicolons and create tokens
named ';'". But it means discard the matching text. Oops.

This commit forces rule names to be quoted (either 'xxx' or "xxx"), so these are
all valid:

  - "-"
  ; ";"
  [ \t\n\r] ;

which will produce tokens name "-" and ";", but which ignores whitespace.