Add a simple static compiler.
This allows lrlex to be embedded (with, admittedly, a run-time dependency on
lrlex) directly into a project. As such, this isn't hugely useful at the moment,
as lrlex is very quick at slurping in .l files. However, it's often not
ergonomic to carry around a .l file, so compiling into the binary is often the
right thing to do (and also matches the conventional behaviour of lex).