Move to a "you have to specify the path of the .l file to be compiled" model.
Since we want lrlex to interface with external libraries (notably, but not
exclusively, lrpar), we need a way to specify the mapping of token IDs between
the two libraries. This needs to be done on a per .l basis, so we can't
generally say "scan all the .l files in this repo" because there's then no sane
way to link up the token ID mapping to each file.

This commit thus forces users to specify the sub-path that .l file lives in the
src/ directory, giving them the opportunity to specify the tokens map. At the
moment we don't do anything with that, but we will do shortly.