Make use of the rule_ids_map parameter and map x.l files to x_l.rs.
The rule_ids_map parameter allows us, at compile-time, to generate constants for
token IDs.

At the same time, map x.l files to x_l.rs. Previously we mapped x.l to x.rs; but
then what should x.y (i.e. the corresponding Yacc file) map to? By mapping x.l
to x_l.rs (and x.y to x_y.rs) we solve this problem neatly. It's also now clear
to me that the previous "look at the timestamp of x.l and x.rs and only save to
x.rs if it is older than x.l" check is incorrect, because someone could (without
realising) change rule IDs on every build. So we now create the output as a
string, and only save it if it's different than the output file's current
contents. This clearly incurs more overhead, but it means that we only force a
recompile when necessary.