A basic compile-time system for generating parsers statically.
This is similar in use to the compile-time system for lrlex, though the
implementation details are very different. For simplicity, we use serde to
serialize the grammar, state graph, and state table at compile-time, loading
them back in later.

There is one major missing feature which is that there is no way to get NDIdxs.
We don't currently provide access to the Grammar struct, nor do we provide
compile-time constants. I need to think more about how to do this, possibly
changing the way that NTIdx is defined in cfgrammar. However, that can be done
later: the bare bones here are, I think, sufficient to see clearly the direction
of travel.