Start moving to a model where users can specify the size of the *Idx newtypes.
In cfgrammar, we have various newtypes for things like production indexes and so
on. The problem is that we have no idea how big an integer type we need for any
given grammar: maybe a u8 can store everything the user will throw at us, but
maybe we need a much bigger type. The safe option is therefore to choose a "big"
integer type and forget about the problem. Unfortunately, one of the index types
-- NTIdx (Non-Terminal index) is used in parse trees. We can end up with
squillions of Nodes after a parse, and the size of the integer type used can be
a noticeable performance penalty (from memory, at least 5-10%).

I fudged this before by making the API for such newtypes expose a u32, but
internally storing a u16. This seemed like a good idea at the time. However, "if
someone gives us a grammar which exceeds a u16, we can transparently change the
storage type to u32" is not a good way of doing things: we'd have had to slow
*everyone's* program down as soon as an example of such a grammar surfaced.

This commit starts to move us to a model where the *Idx newtypes have a
user-controlled type. It defaults to u32, so that users who don't care about
performance are highly unlikely to ever get run-time errors. Users who do care
about performance can choose much smaller types (for all the grammars I know of,
u16 is probably enough; and u8 will often do).

Perhaps surprisingly given the number of different lines touched, this is the
minimal commit which compiles and passes tests: it only changes the PIdx newtype
(which is the least frequently used), using a temporary builder macro in
u32struct.rs. Of course, this still requires that most of the major changes are
made, but it keeps the size of this commit down a bit; subsequent commits will
change the other newtypes.