Document, and sort-of enforce, the constraints on integers we expect.
Rust imposes surprisingly few constraints on integers. That makes it very hard
to deal with code that uses integers of different sizes. For example, u8 and u16
can be converted infallibly to usize, but u32 and bigger can only be fallibly
converted (in nightly) to usize (i.e. using try_from). That means that usize
can't be converted infallibly to u32 or bigger, so code which assumes safety
between u32 and usize is not guaranteed to be portable.

I have tried 4 or 5 of ways of working around this, and nothing has been
satisfactory. This commit enforces the following constraints on the integer
types that cfgrammar will work with:

  * StorageT can be infallibly converted to usize
  * usize can be infallibly converted to StorageT

if either direction were to lead to a loss of precision, a panic occurs. [It
would be nice to have this happen at compile-time, but I haven't been able to
get that to work properly.]

Is this perfect? Clearly not. It means that our code is less portable than
normal Rust code if StorageT > u16. However, the user is in control of this, and
it means we can simplify our code base and reduce the chances that we're bitten
by nasty errors.