Enable "as" conversion from usize to StorageT.
When we're iterating over TIdx/NTIdx values, it's easiest to use a usize
counter. However, since sizeof(StorageT) <= sizeof(usize) we were previously
forced to check that the usize value could fit in a StorageT before doing so.
This is pointless, because by definition we're iterating from 0 to a value that
was a StorageT, so these checks are pointless. What we really want to do is
write something like "usizevalue as storagetvalue", but our use of generics
previously prevented that.

This commit therefore adds a new constraint which is that usize must be
convertible via "as" to StorageT. This requires that StorageT is 'static (which
is OK, because it's only ever going to be a primitive unsigned integer), hence a
lot of (boring) changes needing to be threaded through.