Remove TokId as a generic type.
We previously (sometimes) separated out the notions of "the storage type for
token IDs" from "the storage type for nonterminals and productions." This is too
fine a distinction, in my opinion, for anyone to care about.

This commit thus flattens these into one type, consistently called StorageT. In
other words, StorageT has to be big enough to index a grammar's tokens,
nonterminals, and productions. In most cases, u16 will be more than big enough
to handle this, but (if StorageT is not specified) we default to u32 to be on
the safe side.