Move terms_len to return TIdx<StorageT>.
This has a fairly big knock-on effect, which I've chosen to do in one fell
swoop, notably refactoring a number of issues surrounding StIdx. Those latter
aren't complete in this commit, but it's already fairly clear that the new
approach is simplifying and sanitising code: a number of who-knows-what-they-do
"as" conversions have disappeared in this commit.