Move the astar/dijkstra functions from operating on u32 to u16.
Practically speaking this doesn't make very much difference: the typical max
value used is in the single digits, so u16 has plenty of space.

However, this change is somewhat forced on us, because Rust doesn't guarantee
infallible conversions from u32 to usize (but does guarantee them from u16 to
usize). In order to get rid of the various "as" conversions in astar.rs, we
could either 1) have used lots of try_from conversions (which are ugly and,
in critical loops, not cost free) 2) convert the relevant u32s to u16s. This
commit does the latter.