Remove the biggest crutch: StorageT::from(usize).
This has been hiding a large number of integer casts; it's always safe, in the
sense that that StorageT::from(usize) [OK, the function isn't called that: it's
NTIdx::from(t: <usize>) etc.] dynamically checked that the value being passed
can be stored in StorageT. However, this is lazy: it's unclear to the reader of
the code whether there's a possibility or not that this dynamic check could
fail.

This commit first removes StoargeT::from(usize), then changes lots of code along
the line of NTIdx::from(t) to NTIdx(t.as_()) where "as_" converts a usize to a
StorageT (at least conceptually) without any checks. This required manually
auditing that each such conversion is safe (fortunately the verification was
easy in pretty much every case).

That means we now have faster code that is less obviously incorrect.