Allow users to customise whether missing terminals cause a panic or not.
In general, terminals used in the grammar but not the lexer or bad (so, by
defaut, they cause a panic), whereas terminals defined in the lexer but not
used in the grammar are OK (and hence don't cause a panic by default). However,
we provide a simple switch for both cases to allow users to change the
behaviour.