Move the firsts module to cfgrammar.
The "firsts" calculation has nothing to do specifically with lrtable: it's a
common calculation on grammars. This commit moves it to cfgrammar and gives it
it's own trait (so that other non-Yacc grammar types can also expose such
functionality). However, since this means we end up having to use dynamic
dispatch on the trait, we give YaccGrammar a new method "yacc_firsts" which
returns the YaccFirsts implementation, allowing static dispatch. Thus there's no
performance overhead for this from lrtable's perspective.