Rename prod_firsts to firsts.
The "prod_" part of this is incorrect: we don't store FIRSTS for productions,
but for non-terminals. Which then raises the question of why we're even saying
we're storing FIRSTS for productions/nonterminals, when we can only store them
for nonterminals. So correct the terminology in comments to "nonterminals", and
simplify "prod_firsts" to simply "firsts".

[I suspect this mistake dates form the early days of cfgrammar when I couldn't
remember the proper terms for various parsing constructs. Mea culpa!]