Add a simple panic mode recoverer.
This recoverer is based on that in Compiler Design in C by Allen I. Holub p.348.
See the comments in src/lib/panic.rs for why this isn't a particularly great fit
for lrpar (though it's an interesting point of comparison relative to more
complex algorithms like CPCT+ and MF).