The recoverer needs to be part of the rebuild cache.
The rebuild cache is a very useful performance feature, because it stops us
spending time generating grammar information and having rustc compile it (the
latter is much slower than the former). However, previously we were a bit
sloppy, not included the recovered as part of the rebuild cache.

This commit fixes that oversight, and also documents a little more clearly what
expectations the rebuild cache needs to satisfy.