Rethink the terminology we want to use.
Whatever we choose, we're going to be out of sync with some grammar types, and I
have tried -- it seems to me -- every possible naming convention in my head
before settling on this one. I have chosen it in part because "terminal" and
"nonterminal" are easily confused terms, particularly in grammars, where it's
not really obvious why a "nonterminal" might be called that: "token" and "rule"
are easier to think of IMHO. We can then leave "terminal" and "nonterminal" for
parse trees and the like.