Semi-mechanically rename terminal to token with srep.
This isn't fully mechanical for two reasons. First, there's one instance (in a
comment) where "terminal" was correctly referenced and needed to stay. Second,
there's a clash in a test name. The test had a not-very-good name before, so
renaming it to test_redefine_rules_as_tokens is no bad thing. It was also marked
(weirdly) as #[should_panic], even though it could be changed into a normal test
(which I've also done).