Cherry pick output from rustfmt.
Unfortunately rustfmt doesn't quite do the job we need of it. It has a bug or
two which makes it generate incorrect output in a couple of cases; it struggles
to format a handful of inline comments correctly (which is a bit weird); and it
makes some of the tests much, much harder to read (which isn't really its fault:
it just can't know that formatting things in a different way is the right thing
to do).

What I've done is manually select as much of the formatting it's done as
possible. I've probably selected about 95% of the non-test changes; but quite a
lot of the tests I've left as-is, because they're easier to read as they are.