Force lrpar dependencies to be rebuilt if lrpar is recompiled.
Previously, if you changed the source code of lrpar, then libraries depending on
it wouldn't be recompiled. This caused me some surprise and pain recently when I
changed lrpar and couldn't work out why other libraries were not doing the right
thing.

This commit uses vergen to embed the timestamp of the build into the lrpar
cache; if lrpar is recompiled (e.g. as a result of someone changing the source
code), then the cached timestamp will change, and libraries depending on lrpar
will be recompiled.