Rename the "lrpar" binary to "nimbleparse".
This is forced on us by an upcoming change which -- solely due to the lrpar
binary (i.e. not the lrpar library) -- creates a circular dependency between
lrpar and lrlex. We could resolve this if we could specify different
dependencies for libs and bins in Cargo.toml files, but that is currently
unimplemented (see https://github.com/rust-lang/cargo/issues/1982). So, bite the
bullet and make a top-level binary directory called nimbleparse ("quickparse"
was my original idea, but that suggests it's super speedy, whereas I wanted to
convey "allows you to easily see if an arbitrary grammar can parse an arbitrary
input").