Move the Lexeme type from lrlex to lrpar.
Minimally move the Lexeme type from lrlex to lrpar: in order that people can use
non-lrlex lexers with lrpar, we have to have lrpar specify how lexers should
interact with it, which means that the relevant lexing types (etc.) have to be
in lrpar. This is the first step in that direction.

However, this is a bit more complex than it seems because lrpar relies on lrlex
for some of its tests. However, simply importing lrlex from lrpar would create a
circular dependency, which Rust can't handle. We thus implement a "small lexer"
in lrpar's tests which minimally emulates the relevant parts of lrlex -- enough
to run the tests successfully, at least.