Create a Lexer trait in lrpar.
The idea here is that we can completely decouple lrpar from lrlex (i.e. allow
other lexers to be used with lrpar). Any lexer which wants to be used by lrpar
has to implement the Lexer trait, and generate instances of lrpar's Lexeme type,
but lrpar otherwise makes few demands on it.

This version of the Lexer trait is a minimal port of lrlex's previous interface
(though it has several knock-on effects, notably requiring more type bounds,
that make the commit longer than one would like. This isn't the interface we
want going forwards, but this is very much a "one thing at a time" commit.