Make the Lexer trait into a pseudo-iterator.
In most cases, what we really want to do is to iterate over lexemes one at a
time: for very large files, bundling them up into a Vec first is probably not
very efficient (and it precludes entirely use cases where input is received
incrementally).

This commit makes the Lexer trait be a sort-of iterator (it's not an instance of
Rust's Iterator type, but it's deliberately similar): calling "next" must
produce the next lexeme in the stream. "all_lexemes" is a convenience method
which bundles all lexemes up into a Vec. As Lexer goes, it can optionally record
line number information that can be later queried.