Use a Vec<u32> instead of a HashMap for actions
This replaces the HashMap representation for statetable lookups with a Vector,
encoding actions as a `u32` integer, where the first 2 bits describe the action
type, and the last 30 bits the the value of the action. For example, the action
Shift(3) would be encoded as `0b1101`, and Reduce(4) would be `0b10010`.