Remove the not-obviously-safe conversion from usize to StIdx.
We should never have made the guarantee that you can infallibly convert a usize
into a StIdx: it's clearly not safe in general, and thus each person who makes
such a conversion should think through the consequences carefully.

This PR therefore removes the From conversion from usize to StIdx, pushing it
out to "as" casts in the normal code. Fortunately, these are all easy to verify
as being safe.