Remove the Option type from StateTable::action's return.
This was the result of an incomplete migration of code with the recent
introduction of Action::Error. We ended up with code duplication:

  match action(...) {
    ...
    Some(Action::Error) => XYZ,
    None => XYZ
  }

which is a disaster waiting to happen. Worse, panic mode had been subtly broken
by this change (getting into a loop that it only got out of because of the error
recovery time out).

This commit simply completes the job, changing:

  pub fn action(&self, stidx: StIdx, tidx: TIdx<StorageT>) -> Option<Action<StorageT>>

to:

  pub fn action(&self, stidx: StIdx, tidx: TIdx<StorageT>) -> Action<StorageT>