Skipping the top element in the parsing stack is dangerous.
On the first iteration of panic mode, skipping the top element of the stack is a
valid optimisation: however, if the input is advanced, then the stack is
restored, and the top element of the stack *might* match. On the big corpus of
Java programs, this makes a noticeable (positive) 2% change in the number of
files which can be successfully recovered from.