Use sparsevec for parsing tables
This commit changes the datastructure for storing the actions and
goto tables from PackedVec to SparseVec. This new data structure
first compresses the tables using row displacement and then stores
the result in a PackedVec.

For the Java parser, this results in an 8% smaller binary, saves 67% of memory
for the actions table and 83% for the gotos table, and gives a speed
improvement of 2% when parsing the Java Standard Library.