Don't skip all whitespace equally.
If we have a declaration like:

  %token A

we want the "A" to be on the same line as the "%token". Previously we skipped
all whitespace equally -- including newlines. Thus this was semantically
equivalent to the above:

  %token
  A

This commit splits whitespace skipping into two: skip all whitespace; or skip
all whitespace except newlines. Incorrect input such as the latter now causes an
error:

  Reached end of line without finding expected content at line 1 column 7