Don't interpret comments in %epp declarations as comments.
The Lua grammar has these two lines, one after another:

  %epp FSFS //
  %epp EQEQ ==

The "//" was seen as a comment, so the parser saw this as equivalent to:

  %epp FSFS
  %epp EQEQ ==

So the EPP for FSFS was "%epp EQEQ ==" which, to put it mildly, confused me.

This commit changes things so that the format of an %epp line is:

  %epp<whitespace including comments><token name><whitespace excluding
  comments><all other content>

This feels a bit evil, but it allows the above example to work. The other thing
we could do is use quoting and allow escaping e.g.:

  %epp FSFS "//"
  %epp EQEQ "=="
  %epp STRING "\"...\""

This is more principled, but also more annoying to use...